make every symbol uint64_t in bindings :)

+++ src/libs/cabac/bindings.cpp
@@ -148,42 +148,42 @@ PYBIND11_MODULE(cabac, m) {
     // Context IDs
     m.def("getContextIdBinsOrder1TU", &contextSelector::getContextIdBinsOrder1TU, 
         py::arg("n"), py::arg("symbolPrev"), py::arg("restPos") = 10);
-    m.def("getContextIdsBinsOrder1TU", [](unsigned int symbolPrev, unsigned int restPos = 10, unsigned int numMaxBins = 512) {
+    m.def("getContextIdsBinsOrder1TU", [](uint64_t symbolPrev, unsigned int restPos = 10, unsigned int numMaxBins = 512) {
         std::vector<unsigned int> ctxIDs(numMaxBins, 0);
         contextSelector::getContextIdsBinsOrder1TU(ctxIDs, symbolPrev, restPos);
         return ctxIDs;
     });
     m.def("getContextIdBinsOrder1EG0", &contextSelector::getContextIdBinsOrder1EG0, 
         py::arg("n"), py::arg("symbolPrev"), py::arg("restPos") = 10);
-    m.def("getContextIdsBinsOrder1EG0", [](unsigned int symbolPrev, unsigned int restPos = 10, unsigned int numMaxPrefixBins = 24) {
+    m.def("getContextIdsBinsOrder1EG0", [](uint64_t symbolPrev, unsigned int restPos = 10, unsigned int numMaxPrefixBins = 24) {
         std::vector<unsigned int> ctxIDs(numMaxPrefixBins, 0);
         contextSelector::getContextIdsBinsOrder1EG0(ctxIDs, symbolPrev, restPos);
         return ctxIDs;
     });
     m.def("getContextIdBinsOrder1EGk", &contextSelector::getContextIdBinsOrder1EGk, 
         py::arg("n"), py::arg("symbolPrev"), py::arg("k"), py::arg("restPos") = 10);
-    m.def("getContextIdsBinsOrder1EGk", [](unsigned int symbolPrev, unsigned int k, unsigned int restPos = 10, unsigned int numMaxPrefixBins = 24) {
+    m.def("getContextIdsBinsOrder1EGk", [](uint64_t symbolPrev, unsigned int k, unsigned int restPos = 10, unsigned int numMaxPrefixBins = 24) {
         std::vector<unsigned int> ctxIDs(numMaxPrefixBins, 0);
         contextSelector::getContextIdsBinsOrder1EGk(ctxIDs, symbolPrev, k, restPos);
         return ctxIDs;
     });
     m.def("getContextIdSymbolOrder1TU", &contextSelector::getContextIdSymbolOrder1TU, 
         py::arg("n"), py::arg("symbolPrev"), py::arg("restPos") = 8, py::arg("symbolMax")=32);
-    m.def("getContextIdsSymbolOrder1TU", [](unsigned int symbolPrev, unsigned int restPos = 8, unsigned int symbolMax=32, unsigned int numMaxBins = 512) {
+    m.def("getContextIdsSymbolOrder1TU", [](uint64_t symbolPrev, unsigned int restPos = 8, unsigned int symbolMax=32, unsigned int numMaxBins = 512) {
         std::vector<unsigned int> ctxIDs(numMaxBins, 0);
         contextSelector::getContextIdsSymbolOrder1TU(ctxIDs, symbolPrev, restPos, symbolMax);
         return ctxIDs;
     });
     m.def("getContextIdSymbolOrder1EG0", &contextSelector::getContextIdSymbolOrder1EG0, 
         py::arg("n"), py::arg("symbolPrev"), py::arg("restPos") = 8, py::arg("symbolMax")=32);
-    m.def("getContextIdsSymbolOrder1EG0", [](unsigned int symbolPrev, unsigned int restPos = 8, unsigned int symbolMax=32, unsigned int numMaxPrefixBins = 24) {
+    m.def("getContextIdsSymbolOrder1EG0", [](uint64_t symbolPrev, unsigned int restPos = 8, unsigned int symbolMax=32, unsigned int numMaxPrefixBins = 24) {
         std::vector<unsigned int> ctxIDs(numMaxPrefixBins, 0);
         contextSelector::getContextIdsSymbolOrder1EG0(ctxIDs, symbolPrev, restPos, symbolMax);
         return ctxIDs;
     });
     m.def("getContextIdSymbolOrder1EGk", &contextSelector::getContextIdSymbolOrder1EGk, 
         py::arg("n"), py::arg("symbolPrev"), py::arg("k"), py::arg("restPos") = 8, py::arg("symbolMax")=32);
-    m.def("getContextIdsSymbolOrder1EGk", [](unsigned int symbolPrev, unsigned int k, unsigned int restPos = 8, unsigned int symbolMax=32, unsigned int numMaxPrefixBins = 24) {
+    m.def("getContextIdsSymbolOrder1EGk", [](uint64_t symbolPrev, unsigned int k, unsigned int restPos = 8, unsigned int symbolMax=32, unsigned int numMaxPrefixBins = 24) {
         std::vector<unsigned int> ctxIDs(numMaxPrefixBins, 0);
         contextSelector::getContextIdsSymbolOrder1EGk(ctxIDs, symbolPrev, k, restPos, symbolMax);
         return ctxIDs;
